Add AccessibilityFrame and update ValidationMark
